move hooks to application
