+++ feature/links/application/hooks/browser/useWebBrowser.ts
@@ -4,8 +4,8 @@ import type {
   AllowedDomain,
   WebBrowserOptions,
   WebBrowserResult,
-} from "../domain/models/types";
-import { WebBrowserService } from "../service/WebBrowserService";
+} from "../../../domain/models/types";
+import { WebBrowserService } from "../../../service/WebBrowserService";
 
 export const useWebBrowser = () => {
   const [isLoading, setIsLoading] = useState(false);
