+++ feature/links/application/hooks/link/useLinks.ts
@@ -1,7 +1,7 @@
 import useSWR from "swr";
 
-import { type LinkPreview } from "../domain/models/types";
-import { getLinksPreview } from "../service/linkServices";
+import { type LinkPreview } from "../../../domain/models/types";
+import { getLinksPreview } from "../../../service/linkServices";
 
 type Purpose = "top-view" | "swipe";
